chore: move codeSamples under generate
