+++ cmd/quickstart.go
@@ -3,6 +3,7 @@ package cmd
 import (
 	"context"
 	"fmt"
+	genCmd "github.com/speakeasy-api/speakeasy/cmd/generate"
 	"os"
 	"path/filepath"
 	"strings"
@@ -212,7 +213,7 @@ func quickstartExec(ctx context.Context, flags QuickstartFlags) error {
 		return errors.Wrapf(err, "failed to run speakeasy generate")
 	}
 
-	workflow, err := run.NewWorkflow("Workflow", initialTarget, "", genVersion, "", nil, nil, false, !flags.SkipCompile)
+	workflow, err := run.NewWorkflow("Workflow", initialTarget, "", genCmd.GenVersion, "", nil, nil, false, !flags.SkipCompile)
 	if err != nil {
 		return err
 	}
